requirements.txt
autopep8==1.6.0
python-Levenshtein==0.12.2